+++ backend/freelancePM/src/main/java/com/freelance/freelancepm/service/IInvoiceService.java
@@ -1,14 +1,17 @@
 package com.freelance.freelancepm.service;
- 
+
 import com.freelance.freelancepm.dto.InvoiceCreateRequest;
 import com.freelance.freelancepm.dto.InvoiceResponse;
 import com.freelance.freelancepm.dto.InvoiceUpdateRequest;
- 
+
 import java.util.List;
- 
+
 public interface IInvoiceService {
     InvoiceResponse create(InvoiceCreateRequest req);
-    InvoiceResponse update(Long invoiceId, InvoiceUpdateRequest req);
-    InvoiceResponse getById(Long invoiceId);
+
+    InvoiceResponse update(Integer invoiceId, InvoiceUpdateRequest req);
+
+    InvoiceResponse getById(Integer invoiceId);
+
     List<InvoiceResponse> list();
 }
